Backfill the README with instructions and content from the original
lemurheavy's coveralls-ruby repository apparently updated its README
after this fork, and it has a lot of detail in it that's helpful when
trying to get set up. I think it _used_ to rely on the docs being hosted
by coverall's site, but that site has since stopped holding them, and
now points to _your_ readme, which doesn't include much detail.